+++ matfree/lanczos.py
@@ -130,6 +130,69 @@ def alg_tridiag_full_reortho(depth, /, validate_unit_2_norm=False) -> AlgorithmT
     return _Alg(init=init, step=apply, extract=extract, lower_upper=(0, depth + 1))
 
 
+def alg_bidiag_full_reortho(depth, /, matrix_shape, validate_unit_2_norm=False):
+    """Construct an implementation of **bidiagonalisation**.
+
+    Uses pre-allocation. Fully reorthogonalise vectors at every step.
+
+    Works for **arbitrary matrices**. No symmetry required.
+
+    Decompose a matrix into a product of orthogonal-**bidiagonal**-orthogonal matrices.
+    Use this algorithm for approximate **singular value** decompositions.
+    """
+    nrows, ncols = matrix_shape
+    max_depth = min(nrows, ncols) - 1
+    if depth > max_depth or depth < 0:
+        msg1 = f"Depth {depth} exceeds the matrix' dimensions. "
+        msg2 = f"Expected: 0 <= depth <= min(nrows, ncols) - 1 = {max_depth} "
+        msg3 = f"for a matrix with shape {matrix_shape}."
+        raise ValueError(msg1 + msg2 + msg3)
+
+    class State(containers.NamedTuple):
+        i: int
+        Us: Array
+        Vs: Array
+        alphas: Array
+        betas: Array
+        beta: Array
+        vk: Array
+
+    def init(init_vec: Array) -> State:
+        if validate_unit_2_norm:
+            init_vec = _validate_unit_2_norm(init_vec)
+
+        nrows, ncols = matrix_shape
+        alphas = np.zeros((depth + 1,))
+        betas = np.zeros((depth + 1,))
+        Us = np.zeros((depth + 1, nrows))
+        Vs = np.zeros((depth + 1, ncols))
+        v0, _ = _normalise(init_vec)
+        return State(0, Us, Vs, alphas, betas, np.zeros(()), v0)
+
+    def apply(state: State, Av: Callable, vA: Callable) -> State:
+        i, Us, Vs, alphas, betas, beta, vk = state
+        Vs = Vs.at[i].set(vk)
+        betas = betas.at[i].set(beta)
+
+        uk = Av(vk) - beta * Us[i - 1]
+        uk, alpha = _normalise(uk)
+        uk, *_ = _gram_schmidt_classical(uk, Us)  # full reorthogonalisation
+        Us = Us.at[i].set(uk)
+        alphas = alphas.at[i].set(alpha)
+
+        vk = vA(uk) - alpha * vk
+        vk, beta = _normalise(vk)
+        vk, *_ = _gram_schmidt_classical(vk, Vs)  # full reorthogonalisation
+
+        return State(i + 1, Us, Vs, alphas, betas, beta, vk)
+
+    def extract(state: State, /):
+        _, uk_all, vk_all, alphas, betas, beta, vk = state
+        return uk_all.T, (alphas, betas[1:]), vk_all, (beta, vk)
+
+    return _Alg(init=init, step=apply, extract=extract, lower_upper=(0, depth + 1))
+
+
 def _normalise(vec):
     length = linalg.vector_norm(vec)
     return vec / length, length
